factor out ParentRef into its own file
