+++ examples/somersaultecu.py
@@ -32,6 +32,7 @@ from odxtools.odxlink import OdxDocFragment, OdxLinkId, OdxLinkRef
 from odxtools.odxtypes import DataType
 from odxtools.parameters import (CodedConstParameter, MatchingRequestParameter, NrcConstParameter,
                                  ValueParameter)
+from odxtools.parentref import ParentRef
 from odxtools.physicaltype import PhysicalType
 from odxtools.service import DiagService
 from odxtools.singleecujob import ProgCode, SingleEcuJob
@@ -1751,7 +1752,7 @@ somersault_lazy_diaglayer = DiagLayer(
     long_name="Somersault lazy ECU",
     description="<p>Sloppy variant of the somersault ECU (lazy &lt; assiduous)</p>",
     parent_refs=[
-        DiagLayer.ParentRef(  # <- TODO: this is a bit sketchy IMO
+        ParentRef(  # <- TODO: this is a bit sketchy IMO
             parent=OdxLinkRef.from_id(somersault_diaglayer.odx_id),
             ref_type="BASE-VARIANT-REF",
             # this variant does not do backflips
@@ -1802,7 +1803,7 @@ somersault_assiduous_diaglayer = DiagLayer(
         sdgs=[],
     ),
     parent_refs=[
-        DiagLayer.ParentRef(  # <- TODO: this is a bit sketchy IMO
+        ParentRef(  # <- TODO: this is a bit sketchy IMO
             parent=OdxLinkRef.from_id(somersault_diaglayer.odx_id),
             ref_type="BASE-VARIANT-REF",
             # this variant does everything which the base variant does
