+++ odxtools/write_pdx_file.py
@@ -5,9 +5,12 @@ import inspect
 import os
 import time
 import zipfile
-import jinja2
-import odxtools
 from typing import Any, Dict, List, Optional, Tuple
+
+import jinja2
+
+import odxtools
+
 from .odxtypes import bool_to_odxstr
 
 odxdatabase = None
